requirements.txt
google-adk>=0.5.0
duckduckgo-search>=6.0.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
# marketing tools
pytrends>=4.9.0
feedparser>=6.0.0
# memory store — uses stdlib sqlite3, no extra package needed
# paperclip integration
fastapi>=0.110.0
uvicorn>=0.29.0
